ai-service/requirements.txt
fastapi
uvicorn
python-multipart
python-doctr[torch]
torch
torchvision
tf-keras
